+++ src/main/java/com/example/bcsd/BcsdApplication.java
@@ -2,7 +2,9 @@ package com.example.bcsd;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class BcsdApplication {
