+++ src/main/java/com/example/bcsd/DTO/HelloDTO.java
@@ -1,4 +1,4 @@
-package com.example.bcsd;
+package com.example.bcsd.DTO;
 
 import java.util.List;
